pyproject.toml
[project]
name = "awscvgenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "jinja2>=3.1.6",
    "uvicorn>=0.34.0",
    "boto3>=1.35.0",
    "markdown>=3.7",
    "requests>=2.32.3",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    "python-dotenv==1.0.0",
    "httpx>=0.28.1",
    "pdfkit>=1.0.0",
    "weasyprint>=64.1",
]
